attempt parse and solve with worker
